Retorna um tópico, foi feito DetalhersTopicoDto e RespostaDto

+++ src/main/java/lucasferreira/com/github/forum/repository/TopicoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TopicoRepository  extends JpaRepository<Topico, Integer> {
+public interface TopicoRepository  extends JpaRepository<Topico, Long> {
 
     /*@Query("SELECT t FROM Topico t WHERE t.curso.nome = :nomeCurso")
     List<Topico> carregarPorNomeDoCurso(@Param("nomeCurso")(String nomeCurso);
